version bump
This is probably the last time I'm touching this for a while.  It passes
all tests and I've unbroken things enough so unless someone comes back
with an issue or wants a feature I'm done with it.

OpenNMS 16.0.0 supports REST export of JRD files anyway, so right about
the same time I wrote this we got the built-in support :)