changed from AUTOLOAD to static refs in places
This has tradeoffs.  The advantage is that AUTOLOAD is too loose and
makes everything a runtime check instead of compile time.  The
disadvantage is that you can overwrite a defined sub if you aren't
careful, and you need to make sure all your class variables are defined
when you call new() or it doesn't work (which is why Datasource is still
using AUTOLOAD)